Updated current/ to version 3.3.713.
git-svn-id: svn://svn.bearnip.com/ldmud/3.3/current@2282 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/access_check.c
@@ -386,7 +386,7 @@ read_access_file (void)
 
         /* The rest of the line is the message to print.
          */
-        fgets(message, sizeof(message)-1, infp);
+        fgets(message, (int)sizeof(message)-1, infp);
         message[sizeof(message) - 1] = '\0';
 
         /* Check if this rule creates a new class. If yes, allocate
